pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "datadoctor"
version = "0.4.0"
description = "Autonomous Data Inspection Agent"
readme = "README.md"
authors = [
    { name = "dataDoctor Contributors" }
]
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "streamlit>=1.30.0",
    "plotly>=5.18.0",
    "python-dotenv>=1.0.0",
    "pytest>=7.4.0",
    "numpy>=1.26.0",
    "jsonschema>=4.20.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "python-multipart>=0.0.6",
    "supabase>=2.0.0",
    "joblib>=1.3.0",
    "shap>=0.42.0",
]

[project.scripts]
datadoctor = "cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "cli.py"]
